+++ chromeos/ash/services/quick_pair/fast_pair_data_parser_unittest.cc
@@ -660,7 +660,7 @@ TEST_F(FastPairDataParserTest, ParseMessageStreamMessage_EnableSilenceMode) {
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 1);
+        EXPECT_EQ(messages.size(), 1u);
         EXPECT_TRUE(messages[0]->is_enable_silence_mode());
         EXPECT_TRUE(messages[0]->get_enable_silence_mode());
         run_loop.Quit();
@@ -682,7 +682,7 @@ TEST_F(FastPairDataParserTest,
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 0);
+        EXPECT_EQ(messages.size(), 0u);
         run_loop.Quit();
       });
 
@@ -699,7 +699,7 @@ TEST_F(FastPairDataParserTest, ParseMessageStreamMessage_DisableSilenceMode) {
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 1);
+        EXPECT_EQ(messages.size(), 1u);
         EXPECT_TRUE(messages[0]->is_enable_silence_mode());
         EXPECT_FALSE(messages[0]->get_enable_silence_mode());
         run_loop.Quit();
@@ -719,7 +719,7 @@ TEST_F(FastPairDataParserTest,
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 0);
+        EXPECT_EQ(messages.size(), 0u);
         run_loop.Quit();
       });
 
@@ -737,7 +737,7 @@ TEST_F(FastPairDataParserTest,
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 1);
+        EXPECT_EQ(messages.size(), 1u);
         EXPECT_TRUE(messages[0]->is_companion_app_log_buffer_full());
         EXPECT_TRUE(messages[0]->get_companion_app_log_buffer_full());
         run_loop.Quit();
@@ -757,7 +757,7 @@ TEST_F(FastPairDataParserTest,
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 0);
+        EXPECT_EQ(messages.size(), 0u);
         run_loop.Quit();
       });
 
@@ -777,7 +777,7 @@ TEST_F(FastPairDataParserTest,
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 0);
+        EXPECT_EQ(messages.size(), 0u);
         run_loop.Quit();
       });
 
@@ -796,7 +796,7 @@ TEST_F(FastPairDataParserTest, ParseMessageStreamMessage_ModelId) {
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 1);
+        EXPECT_EQ(messages.size(), 1u);
         EXPECT_TRUE(messages[0]->is_model_id());
         EXPECT_EQ(messages[0]->get_model_id(), "AABBCC");
         run_loop.Quit();
@@ -818,7 +818,7 @@ TEST_F(FastPairDataParserTest, ParseMessageStreamMessage_BleAddress) {
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 1);
+        EXPECT_EQ(messages.size(), 1u);
         EXPECT_TRUE(messages[0]->is_ble_address_update());
         EXPECT_EQ(messages[0]->get_ble_address_update(), "AA:BB:CC:DD:EE:FF");
         run_loop.Quit();
@@ -841,7 +841,7 @@ TEST_F(FastPairDataParserTest,
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 0);
+        EXPECT_EQ(messages.size(), 0u);
         run_loop.Quit();
       });
 
@@ -860,7 +860,7 @@ TEST_F(FastPairDataParserTest, ParseMessageStreamMessage_BatteryNotification) {
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 1);
+        EXPECT_EQ(messages.size(), 1u);
         EXPECT_TRUE(messages[0]->is_battery_update());
         EXPECT_EQ(messages[0]->get_battery_update()->left_bud_info->percentage,
                   87);
@@ -885,7 +885,7 @@ TEST_F(FastPairDataParserTest, ParseMessageStreamMessage_RemainingBatteryTime) {
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 1);
+        EXPECT_EQ(messages.size(), 1u);
         EXPECT_TRUE(messages[0]->is_remaining_battery_time());
         EXPECT_EQ(messages[0]->get_remaining_battery_time(), 240);
         run_loop.Quit();
@@ -907,7 +907,7 @@ TEST_F(FastPairDataParserTest,
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 1);
+        EXPECT_EQ(messages.size(), 1u);
         EXPECT_TRUE(messages[0]->is_remaining_battery_time());
         EXPECT_EQ(messages[0]->get_remaining_battery_time(), 271);
         run_loop.Quit();
@@ -927,7 +927,7 @@ TEST_F(FastPairDataParserTest,
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 0);
+        EXPECT_EQ(messages.size(), 0u);
         run_loop.Quit();
       });
 
@@ -945,7 +945,7 @@ TEST_F(FastPairDataParserTest, ParseMessageStreamMessage_ModelIdInvalidLength) {
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 0);
+        EXPECT_EQ(messages.size(), 0u);
         run_loop.Quit();
       });
 
@@ -964,7 +964,7 @@ TEST_F(FastPairDataParserTest,
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 0);
+        EXPECT_EQ(messages.size(), 0u);
         run_loop.Quit();
       });
 
@@ -983,7 +983,7 @@ TEST_F(FastPairDataParserTest,
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 0);
+        EXPECT_EQ(messages.size(), 0u);
         run_loop.Quit();
       });
 
@@ -1002,7 +1002,7 @@ TEST_F(FastPairDataParserTest,
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 0);
+        EXPECT_EQ(messages.size(), 0u);
         run_loop.Quit();
       });
 
@@ -1020,7 +1020,7 @@ TEST_F(FastPairDataParserTest,
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 0);
+        EXPECT_EQ(messages.size(), 0u);
         run_loop.Quit();
       });
 
@@ -1038,7 +1038,7 @@ TEST_F(FastPairDataParserTest, ParseMessageStreamMessage_ActiveComponents) {
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 1);
+        EXPECT_EQ(messages.size(), 1u);
         EXPECT_TRUE(messages[0]->is_active_components_byte());
         EXPECT_EQ(messages[0]->get_active_components_byte(), 0x03);
         run_loop.Quit();
@@ -1058,7 +1058,7 @@ TEST_F(FastPairDataParserTest, ParseMessageStreamMessage_AndroidPlatform) {
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 1);
+        EXPECT_EQ(messages.size(), 1u);
         EXPECT_TRUE(messages[0]->is_sdk_version());
         EXPECT_EQ(messages[0]->get_sdk_version(), 28);
         run_loop.Quit();
@@ -1078,7 +1078,7 @@ TEST_F(FastPairDataParserTest,
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 0);
+        EXPECT_EQ(messages.size(), 0u);
         run_loop.Quit();
       });
 
@@ -1097,7 +1097,7 @@ TEST_F(FastPairDataParserTest, ParseMessageStreamMessage_InvalidPlatform) {
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 0);
+        EXPECT_EQ(messages.size(), 0u);
         run_loop.Quit();
       });
 
@@ -1115,7 +1115,7 @@ TEST_F(FastPairDataParserTest, ParseMessageStreamMessage_RingDeviceNoTimeout) {
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 1);
+        EXPECT_EQ(messages.size(), 1u);
         EXPECT_TRUE(messages[0]->is_ring_device_event());
         EXPECT_EQ(messages[0]->get_ring_device_event()->ring_device_byte, 0x01);
         EXPECT_EQ(messages[0]->get_ring_device_event()->timeout_in_seconds, -1);
@@ -1136,7 +1136,7 @@ TEST_F(FastPairDataParserTest, ParseMessageStreamMessage_RingDeviceTimeout) {
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 1);
+        EXPECT_EQ(messages.size(), 1u);
         EXPECT_TRUE(messages[0]->is_ring_device_event());
         EXPECT_EQ(messages[0]->get_ring_device_event()->ring_device_byte, 0x01);
         EXPECT_EQ(messages[0]->get_ring_device_event()->timeout_in_seconds, 60);
@@ -1157,7 +1157,7 @@ TEST_F(FastPairDataParserTest, ParseMessageStreamMessage_RingInvalidLength) {
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 0);
+        EXPECT_EQ(messages.size(), 0u);
         run_loop.Quit();
       });
 
@@ -1176,7 +1176,7 @@ TEST_F(FastPairDataParserTest,
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 0);
+        EXPECT_EQ(messages.size(), 0u);
         run_loop.Quit();
       });
 
@@ -1194,7 +1194,7 @@ TEST_F(FastPairDataParserTest, ParseMessageStreamMessage_Ack) {
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 1);
+        EXPECT_EQ(messages.size(), 1u);
         EXPECT_TRUE(messages[0]->is_acknowledgement());
         EXPECT_EQ(messages[0]->get_acknowledgement()->action_message_code,
                   0x01);
@@ -1219,7 +1219,7 @@ TEST_F(FastPairDataParserTest, ParseMessageStreamMessage_Nak) {
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 1);
+        EXPECT_EQ(messages.size(), 1u);
         EXPECT_TRUE(messages[0]->is_acknowledgement());
         EXPECT_EQ(messages[0]->get_acknowledgement()->action_message_code,
                   0x01);
@@ -1245,7 +1245,7 @@ TEST_F(FastPairDataParserTest,
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 0);
+        EXPECT_EQ(messages.size(), 0u);
         run_loop.Quit();
       });
 
@@ -1265,7 +1265,7 @@ TEST_F(FastPairDataParserTest, ParseMessageStreamMessage_AckInvalidLength) {
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 0);
+        EXPECT_EQ(messages.size(), 0u);
         run_loop.Quit();
       });
 
@@ -1283,7 +1283,7 @@ TEST_F(FastPairDataParserTest, ParseMessageStreamMessage_NakInvalidLength) {
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 0);
+        EXPECT_EQ(messages.size(), 0u);
         run_loop.Quit();
       });
 
@@ -1296,7 +1296,7 @@ TEST_F(FastPairDataParserTest, ParseMessageStreamMessage_NotEnoughBytes) {
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 0);
+        EXPECT_EQ(messages.size(), 0u);
         run_loop.Quit();
       });
 
@@ -1322,7 +1322,7 @@ TEST_F(FastPairDataParserTest,
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 2);
+        EXPECT_EQ(messages.size(), 2u);
         EXPECT_TRUE(messages[0]->is_ring_device_event());
         EXPECT_EQ(messages[0]->get_ring_device_event()->ring_device_byte, 0x01);
         EXPECT_EQ(messages[0]->get_ring_device_event()->timeout_in_seconds, -1);
@@ -1353,7 +1353,7 @@ TEST_F(FastPairDataParserTest,
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 1);
+        EXPECT_EQ(messages.size(), 1u);
         EXPECT_TRUE(messages[0]->is_ring_device_event());
         EXPECT_EQ(messages[0]->get_ring_device_event()->ring_device_byte, 0x01);
         EXPECT_EQ(messages[0]->get_ring_device_event()->timeout_in_seconds, -1);
@@ -1381,7 +1381,7 @@ TEST_F(FastPairDataParserTest,
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 0);
+        EXPECT_EQ(messages.size(), 0u);
         run_loop.Quit();
       });
 
@@ -1409,7 +1409,7 @@ TEST_F(FastPairDataParserTest,
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 1);
+        EXPECT_EQ(messages.size(), 1u);
         EXPECT_TRUE(messages[0]->is_battery_update());
         EXPECT_EQ(87,
                   messages[0]->get_battery_update()->left_bud_info->percentage);
@@ -1443,7 +1443,7 @@ TEST_F(FastPairDataParserTest,
   base::RunLoop run_loop;
   auto callback = base::BindLambdaForTesting(
       [&run_loop](std::vector<mojom::MessageStreamMessagePtr> messages) {
-        EXPECT_EQ(static_cast<int>(messages.size()), 1);
+        EXPECT_EQ(messages.size(), 1u);
         EXPECT_TRUE(messages[0]->is_battery_update());
         EXPECT_EQ(87,
                   messages[0]->get_battery_update()->left_bud_info->percentage);
