+++ components/commerce/core/subscriptions/subscriptions_server_proxy_unittest.cc
@@ -178,7 +178,7 @@ TEST_F(SubscriptionsServerProxyTest, TestCreate) {
           [](base::RunLoop* run_loop, SubscriptionsRequestStatus status,
              std::unique_ptr<std::vector<CommerceSubscription>> subscriptions) {
             ASSERT_EQ(SubscriptionsRequestStatus::kSuccess, status);
-            ASSERT_EQ(1, static_cast<int>(subscriptions->size()));
+            ASSERT_EQ(1u, subscriptions->size());
             auto subscription = (*subscriptions)[0];
             ASSERT_EQ(SubscriptionType::kPriceTrack, subscription.type);
             ASSERT_EQ(IdentifierType::kProductClusterId, subscription.id_type);
@@ -220,7 +220,7 @@ TEST_F(SubscriptionsServerProxyTest, TestCreate_ServerFailed) {
           [](base::RunLoop* run_loop, SubscriptionsRequestStatus status,
              std::unique_ptr<std::vector<CommerceSubscription>> subscriptions) {
             ASSERT_EQ(SubscriptionsRequestStatus::kServerInternalError, status);
-            ASSERT_EQ(0, static_cast<int>(subscriptions->size()));
+            ASSERT_EQ(0u, subscriptions->size());
             run_loop->Quit();
           },
           &run_loop));
@@ -242,7 +242,7 @@ TEST_F(SubscriptionsServerProxyTest, TestCreate_WrongHttpCode) {
           [](base::RunLoop* run_loop, SubscriptionsRequestStatus status,
              std::unique_ptr<std::vector<CommerceSubscription>> subscriptions) {
             ASSERT_EQ(SubscriptionsRequestStatus::kServerParseError, status);
-            ASSERT_EQ(0, static_cast<int>(subscriptions->size()));
+            ASSERT_EQ(0u, subscriptions->size());
             run_loop->Quit();
           },
           &run_loop));
@@ -264,7 +264,7 @@ TEST_F(SubscriptionsServerProxyTest, TestCreate_EmptyResponse) {
           [](base::RunLoop* run_loop, SubscriptionsRequestStatus status,
              std::unique_ptr<std::vector<CommerceSubscription>> subscriptions) {
             ASSERT_EQ(SubscriptionsRequestStatus::kServerInternalError, status);
-            ASSERT_EQ(0, static_cast<int>(subscriptions->size()));
+            ASSERT_EQ(0u, subscriptions->size());
             run_loop->Quit();
           },
           &run_loop));
@@ -286,7 +286,7 @@ TEST_F(SubscriptionsServerProxyTest, TestDelete) {
           [](base::RunLoop* run_loop, SubscriptionsRequestStatus status,
              std::unique_ptr<std::vector<CommerceSubscription>> subscriptions) {
             ASSERT_EQ(SubscriptionsRequestStatus::kSuccess, status);
-            ASSERT_EQ(1, static_cast<int>(subscriptions->size()));
+            ASSERT_EQ(1u, subscriptions->size());
             auto subscription = (*subscriptions)[0];
             ASSERT_EQ(SubscriptionType::kPriceTrack, subscription.type);
             ASSERT_EQ(IdentifierType::kProductClusterId, subscription.id_type);
@@ -328,7 +328,7 @@ TEST_F(SubscriptionsServerProxyTest, TestDelete_ServerFailed) {
           [](base::RunLoop* run_loop, SubscriptionsRequestStatus status,
              std::unique_ptr<std::vector<CommerceSubscription>> subscriptions) {
             ASSERT_EQ(SubscriptionsRequestStatus::kServerInternalError, status);
-            ASSERT_EQ(0, static_cast<int>(subscriptions->size()));
+            ASSERT_EQ(0u, subscriptions->size());
             run_loop->Quit();
           },
           &run_loop));
@@ -350,7 +350,7 @@ TEST_F(SubscriptionsServerProxyTest, TestGet) {
           [](base::RunLoop* run_loop, SubscriptionsRequestStatus status,
              std::unique_ptr<std::vector<CommerceSubscription>> subscriptions) {
             ASSERT_EQ(SubscriptionsRequestStatus::kSuccess, status);
-            ASSERT_EQ(1, static_cast<int>(subscriptions->size()));
+            ASSERT_EQ(1u, subscriptions->size());
             auto subscription = (*subscriptions)[0];
             ASSERT_EQ(SubscriptionType::kPriceTrack, subscription.type);
             ASSERT_EQ(IdentifierType::kProductClusterId, subscription.id_type);
@@ -376,7 +376,7 @@ TEST_F(SubscriptionsServerProxyTest, TestGet_WrongType) {
           [](base::RunLoop* run_loop, SubscriptionsRequestStatus status,
              std::unique_ptr<std::vector<CommerceSubscription>> subscriptions) {
             ASSERT_EQ(SubscriptionsRequestStatus::kInvalidArgument, status);
-            ASSERT_EQ(0, static_cast<int>(subscriptions->size()));
+            ASSERT_EQ(0u, subscriptions->size());
             run_loop->Quit();
           },
           &run_loop));
@@ -398,7 +398,7 @@ TEST_F(SubscriptionsServerProxyTest, TestGet_WrongHttpCode) {
           [](base::RunLoop* run_loop, SubscriptionsRequestStatus status,
              std::unique_ptr<std::vector<CommerceSubscription>> subscriptions) {
             ASSERT_EQ(SubscriptionsRequestStatus::kServerParseError, status);
-            ASSERT_EQ(0, static_cast<int>(subscriptions->size()));
+            ASSERT_EQ(0u, subscriptions->size());
             run_loop->Quit();
           },
           &run_loop));
@@ -422,7 +422,7 @@ TEST_F(SubscriptionsServerProxyTest, TestGet_FetchError) {
           [](base::RunLoop* run_loop, SubscriptionsRequestStatus status,
              std::unique_ptr<std::vector<CommerceSubscription>> subscriptions) {
             ASSERT_EQ(SubscriptionsRequestStatus::kServerParseError, status);
-            ASSERT_EQ(0, static_cast<int>(subscriptions->size()));
+            ASSERT_EQ(0u, subscriptions->size());
             run_loop->Quit();
           },
           &run_loop));
@@ -444,7 +444,7 @@ TEST_F(SubscriptionsServerProxyTest, TestGet_NoSubscriptions) {
           [](base::RunLoop* run_loop, SubscriptionsRequestStatus status,
              std::unique_ptr<std::vector<CommerceSubscription>> subscriptions) {
             ASSERT_EQ(SubscriptionsRequestStatus::kSuccess, status);
-            ASSERT_EQ(0, static_cast<int>(subscriptions->size()));
+            ASSERT_EQ(0u, subscriptions->size());
             run_loop->Quit();
           },
           &run_loop));
